Move source attribute test to appropriate file
